64-bit optimization: Convert bitmap/texture memory to size_t
Convert bitmap and texture data size operations from int to size_t
to remove the 2GB limit on texture data and enable proper 64-bit
memory allocation for large assets.

Core bitmap class changes:
- cBitmapData::mlSize: int → size_t (Bitmap.h line 39)
- cBitmapData::SetData(): int → size_t parameter (Bitmap.h line 36)
- cBitmap::CreateData(): size_t for data size calculation (Bitmap.cpp line 132)

Bitmap loader updates:
- BitmapLoaderSTB.cpp: size_t for image data size (line 95)
- BitmapLoaderDDS.cpp: Removed unnecessary static_cast (line 141)

Texture/OpenGL integration:
- SDLTexture.h: Updated CopyTextureDataToGL() and GenerateMipMaps()
  to accept size_t data size parameters (lines 88, 92)
- SDLTexture.cpp: Updated implementations and added explicit casts
  to GLsizei for OpenGL API calls (lines 718, 793, 918, 931, 940, 947)

Benefits:
- Removes 2GB limit on individual texture data (previously limited by signed int)
- Enables loading of 8K+ textures and large cubemaps
- Proper 64-bit memory addressing for texture assets
- Consistent with Phase 2/3 size_t conversions

Updated README.md to document bitmap/texture memory optimizations.

+++ HPL2/core/include/graphics/Bitmap.h
@@ -33,10 +33,10 @@ namespace hpl {
 		cBitmapData();
 		~cBitmapData();
 
-		void SetData(const unsigned char* apData, int alSize);
+		void SetData(const unsigned char* apData, size_t alSize);
 
 		unsigned char *mpData;
-		int mlSize;	
+		size_t mlSize;	
 	};
 
 	//-----------------------------------------
